NOJIRA Move DbEnum to shared space. This is being used in the front end for identifying potential flash options
