+++ src-electron/db/query-config.js
@@ -22,7 +22,7 @@
  */
 const dbApi = require('./db-api.js')
 const dbMapping = require('./db-mapping.js')
-const dbEnum = require('./db-enum.js')
+const dbEnum = require('../../src-shared/db-enum.js')
 const queryZcl = require('./query-zcl.js')
 
 /**
