+++ src-electron/generator/helper-zcl.js
@@ -16,7 +16,7 @@
  */
 
 const queryZcl = require('../db/query-zcl.js')
-const dbEnum = require('../db/db-enum.js')
+const dbEnum = require('../../src-shared/db-enum.js')
 const templateUtil = require('./template-util.js')
 
 /**
