+++ src-electron/generator/template-util.js
@@ -16,7 +16,7 @@
  */
 
 const queryPackage = require('../db/query-package.js')
-const dbEnum = require('../db/db-enum.js')
+const dbEnum = require('../../src-shared/db-enum.js')
 
 /**
  * @module JS API: generator logic
