+++ src-electron/main-process/menu.js
@@ -27,7 +27,7 @@ const windowJs = require('./window.js')
 const preference = require('./preference.js')
 const generationEngine = require('../generator/generation-engine.js')
 const queryPackage = require('../db/query-package.js')
-const dbEnum = require('../db/db-enum.js')
+const dbEnum = require('../../src-shared/db-enum.js')
 
 var httpPort
 var handlebarTemplateDirectory = __dirname + '/../../test/gen-template'
