+++ src-electron/rest/generation.js
@@ -23,7 +23,7 @@
 
 const generationEngine = require('../generator/generation-engine.js')
 const queryPackage = require('../db/query-package.js')
-const dbEnum = require('../db/db-enum.js')
+const dbEnum = require('../../src-shared/db-enum.js')
 
 /**
  *
