+++ test/importexport.test.js
@@ -21,7 +21,7 @@ const path = require('path')
 const fs = require('fs')
 const importJs = require('../src-electron/importexport/import.js')
 const exportJs = require('../src-electron/importexport/export.js')
-const dbEnum = require('../src-electron/db/db-enum.js')
+const dbEnum = require('../src-shared/db-enum.js')
 const dbApi = require('../src-electron/db/db-api.js')
 const env = require('../src-electron/util/env.js')
 const zclLoader = require('../src-electron/zcl/zcl-loader.js')
