+++ test/server.test.js
@@ -22,7 +22,7 @@ const fs = require('fs')
 const path = require('path')
 const axios = require('axios')
 const dbApi = require('../src-electron/db/db-api.js')
-const dbEnum = require('../src-electron/db/db-enum.js')
+const dbEnum = require('../src-shared/db-enum.js')
 const queryZcl = require('../src-electron/db/query-zcl.js')
 const queryGeneric = require('../src-electron/db/query-generic.js')
 const queryPackage = require('../src-electron/db/query-package.js')
